dcm-mapping/requirements.txt
attrs==21.4.0
beautifulsoup4==4.10.0
blis==0.7.5
bs4==0.0.1
catalogue==2.0.6
charset-normalizer==2.0.10
cymem==2.0.6
deep-translator==1.7.0
editdistance==0.6.0
et-xmlfile==1.1.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
iniconfig==1.1.1
joblib==1.1.0
langcodes==3.3.0
murmurhash==1.0.6
nltk==3.6.7
openpyxl==3.0.9
pathy==0.6.1
pluggy==1.0.0
preshed==3.0.6
py==1.11.0
pydantic==1.8.2
pytest==6.2.5
python-Levenshtein==0.12.2
regex==2022.1.18
rfc3986==1.5.0
scikit-learn==1.0.2
scipy==1.7.3
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.1
spacy==3.2.1
spacy-langdetect==0.1.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.7.4.3
wasabi==0.9.0
webencodings==0.5.1
adal==1.2.5
alembic==0.9.6
aniso8601==8.0.0
azure-common
azure-core
azure-mgmt-compute
azure-mgmt-core
azure-mgmt-datafactory
azure-storage-blob
azure-storage-file-share
bcrypt==3.1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click
coverage==4.4.2
cryptography==2.9.2
datafactory==1.0.1
enum-compat==0.0.3
eventlet==0.21.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-PyMongo==2.3.0
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
greenlet==0.4.16
gunicorn==19.7.1
idna==2.10
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.11.2
jsonschema==2.6.0
llvmlite==0.34.0
Mako==1.0.7
MarkupSafe==1.1.1
msrest==0.6.17
msrestazure==0.6.4
numba==0.51.2
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
pandas==1.1.1
psycopg2-binary==2.8.5
pyarrow==2.0.0
pycparser==2.18
pydrill==0.3.4
PyExcelerate==0.9.0
PyJWT==1.5.3
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2017.3
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.8.1
six==1.11.0
SQLAlchemy==1.3.18
thrift==0.13.0
urllib3==1.25.9
Werkzeug==0.15.3
xlrd==1.2.0
auto-mapper